+++ example/lib/main.dart
@@ -68,7 +68,7 @@ class _MyAppState extends State<MyApp> {
               GlobalMaterialLocalizations.delegate,
               GlobalWidgetsLocalizations.delegate,
               GlobalCupertinoLocalizations.delegate,
-              CriteriaSelectorLocalizationsDelegate(),
+              SelectorLocalizationsDelegate(),
             ],
             supportedLocales: const [
               Locale('en'),